rename ibchooks and lint
