--- go.work
@@ -1,6 +1,0 @@
-go 1.20
-
-use (
-	.
-	./tests/interchaintest
-)
