+++ x/ibchooks/types/errors.go
@@ -1,6 +1,6 @@
 package types
 
-import sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+import sdkerrors "cosmossdk.io/errors"
 
 var (
 	ErrBadMetadataFormatMsg = "wasm metadata not properly formatted for: '%v'. %s"
